Fix bit-depth assumptions in segmentation to support 16-bit DNG files
- Updated calculate_BG_image in segmentation_mod.py to use the dynamic image dtype instead of hardcoded int16.
- Removed hardcoded uint8 casts in fiber and sheet segmentation modules, replacing them with dynamic dtype selection.

+++ myptv/sheets/sheet_segmentation_mod.py
@@ -100,7 +100,7 @@ class sheet_segmentation(object):
         local_mean = convolve2d(image, window, mode='same')
         new_im = image - local_mean
         new_im[new_im<0] = 0
-        new_im = new_im.astype('uint8')
+        new_im = new_im.astype(self.im.dtype)
         return new_im
